Moving findNativeLibPath function to pkg/android/
